+++ backend/src/controllers/testcases.controller.js
@@ -1,10 +1,13 @@
-import {ApiError} from "../middlewares/ApiError.js";
-import {ApiResponse} from "../middlewares/ApiResponse.js";
+import {ApiError} from "../utilities/ApiError.js";
+import {ApiResponse} from "../utilities/ApiResponse.js";
 import {Testcase} from "../models/testcases.model.js";
 
 export const createTestcase = async (req, res, next) => {
     //creates a single testcase of a single problem
     try {
+        if(!req.user.isAdmin){
+            throw new ApiError(403, "You are not authorized to add testcases");
+        }
         const {taskId, input, output} = req.body;
         const testcase = await Testcase.create({taskId, input, output});
         return res.status(201).json(new ApiResponse(201, testcase));
@@ -27,6 +30,9 @@ export const getTestcases = async (req, res, next) => {
 export const deleteTestcase = async (req, res, next) => {
     //deletes a single testcase of a single problem
     try {
+        if(!req.user.isAdmin){
+            throw new ApiError(403, "You are not authorized to delete testcases");
+        }
         const id = req.query.id;
         const testcase = await Testcase.findByIdAndDelete(id);
         return res.status(200).json(new ApiResponse(200, testcase, "Testcase deleted successfully"));
@@ -38,6 +44,9 @@ export const deleteTestcase = async (req, res, next) => {
 export const updateTestcase = async (req, res, next) => {
     //updates a single testcase of a single problem
     try {
+        if(!req.user.isAdmin){
+            throw new ApiError(403, "You are not authorized to update testcases");
+        }
         const id = req.query.id;
         const {input, output} = req.body;
         const testcase = await Testcase.findByIdAndUpdate(id, {input, output}, {new: true});
